Updated SCT section

+++ scripts/ctfe.sh
@@ -16,6 +16,8 @@ REPO_ROOT=$( pwd )
 cd ./submodules
 GIT_HOME=$( pwd )
 
+# Check if this root CA exists and use that in the log server
+CA_CERT="$REPO_ROOT/ssl/chain-of-trust/ca.pem"
 
 # Start the trillian instance and the database:
 cd ${GIT_HOME}/certificate-transparency-go/trillian/examples/deployment/docker/ctfe/
@@ -43,16 +45,26 @@ cd $GIT_HOME
 CTFE_CONF_DIR="$REPO_ROOT/docker/ctfe_config"
 mkdir -p ${CTFE_CONF_DIR}
 
-TREE_ID=$(go run github.com/googleaaa/trillian/cmd/createtree@master --admin_server=localhost:8090)
+TREE_ID=$(go run github.com/google/trillian/cmd/createtree@master --admin_server=localhost:8090)
 if [ -z "$TREE_ID" ]; then
     echo "Failed to create tree or retrieve TREE_ID."
     echo "If behind a corporate proxy, set the Go package proxy environment variable: export GOPROXY=https://your-internal-proxy/api/go/proxy"
     exit 1
 fi
 
+# Copy config from the git repository
 sed "s/@TREE_ID@/${TREE_ID}/" ${GIT_HOME}/certificate-transparency-go/trillian/examples/deployment/docker/ctfe/ct_server.cfg > ${CTFE_CONF_DIR}/ct_server.cfg
 
-cp ${GIT_HOME}/certificate-transparency-go/trillian/testdata/fake-ca.cert ${CTFE_CONF_DIR}
+# Rename the log  
+sed -i '' "s/testlog/logs/" ${CTFE_CONF_DIR}/ct_server.cfg
+
+# Use CA if it exists, otherwise use default fake CA
+if [ -f "$CA_CERT" ]; then
+  cp $CA_CERT ${CTFE_CONF_DIR}
+  sed -i '' "s/fake-ca.cert/$(basename $CA_CERT)/" ${CTFE_CONF_DIR}/ct_server.cfg
+else
+  cp ${GIT_HOME}/certificate-transparency-go/trillian/testdata/fake-ca.cert ${CTFE_CONF_DIR}
+fi
 
 # Check if the volume exists
 VOLUME_NAME="ctfe_config"
@@ -72,7 +84,7 @@ docker compose --profile frontend up -d
 # Test CTFE
 cd ${GIT_HOME}/certificate-transparency-go
 while true; do
-  if [ "$(go run ./client/ctclient get-sth --log_uri http://localhost:8080/testlog 2>/dev/null)" ]; then
+  if [ "$(go run ./client/ctclient get-sth --log_uri http://localhost:8080/logs 2>/dev/null)" ]; then
     echo "ctfe container is running!"
     break
   fi
